+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG - Neo4jH3
 
+## 5.8 2023-0519
+
+* [Added] Tested through Neo4j 5.7
+* [Added] Added support for converting a LINESTRING geospatial object to H3 number - com.neo4jh3.lineash3
+* [Added] Added support for converting a LINESTRING geospatial object to H3 string - com.neo4jh3.lineash3String
+* [Updated] Bumped H3 library to 4.1.1
+
+
 ## 5.7 2023-05-01
 
 * [Added] Tested through Neo4j versions 5.7